Lint backbone.js using eslint. Get eslint running and the qunit tests running wh en `npm test` is executed.
This will help to make our code a bit more future-proof. I copied over the `.eslintrc` file that we were using in the khan-linter repo and removed the React-specific lines. I then added `eslint` as a dependency and set it to run when `npm test` was run. Finally I removed the dependency on Phantomjs and used the `qunit` test runner to execute the tests (making some minor changes to the tests to get them to run in Node-land and deleting some no-longer-needed test files).

Test Plan:
Run `npm test` and see that the linter and tests pass.

Auditors: jlfwong, emily